Playerスクリプトのファイル整理

+++ Assets/Scripts/Field/UiHp.cs
@@ -9,14 +9,12 @@ namespace Field
         [SerializeField] private HpStatus atamakko;
         private Slider _hpGauge;
         
-        // Start is called before the first frame update
         void Start()
         {
             _hpGauge = gameObject.GetComponent<Slider>();
 
             atamakko.MyHP
-                // ReSharper disable once PossibleLossOfFraction
-                .Subscribe(hp => _hpGauge.value = hp / 6)
+                .Subscribe(hp => _hpGauge.value = (float)hp / 6)
                 .AddTo(this);
         }
     }
